requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database (Async Postgres)
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
# psycopg2-binary==2.9.9  # Not needed - using asyncpg instead

# Caching
redis==5.0.1
pymemcache==4.0.0

# AI/ML Libraries
groq==0.4.0
requests==2.31.0
pillow>=10.0.0  # Use latest stable version

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and utilities
python-json-logger==2.0.7

# Streamlit UI
streamlit>=1.28.0
plotly>=5.18.0
pandas>=2.1.0
